Extract program to separate classes
